viennamini ui now handles contacts (single,range) and models
i.e., mobility and scattering.

with mobility, the dependent hierarchy is respected,
meaning that if one turns on ionized impurity, lattice scattering
is turned on and deactivated to the user -> indicating that this dependency
is a given. if ionizied is turned off again, lattice gets active
and stays turned on.

now, off to ViennaMini to link this API to the backend